+++ ViewModel/Pages/FavouritePlayListViewModel.cs
@@ -60,6 +60,7 @@ namespace WinUIMusicPlayer.ViewModel
         {
             UpdateAlbumMenuOptionsPlayList();
             UpDateUsbDeviceMenuflyout();
+            ToolUtils.UpdateMusicMenuAvailability(MenuOptions, SelectedMusics, SelectedMusic);
         }
 
         public void UpDateUsbDeviceMenuflyout()